Ogives n.1
Simple script which simulates the dynamics of the cell-cargo truck system, referring to the phenomenological assumption claimed in [1].
The analysis of the cargo is performed, focusing on the computation of the the Memory Kernel, a characteristic term which eludes for a non - Markovian dynamics of the ensemble.
The algorithm is obtain by discretization of a Volterra Eq. of the 2nd kind, referring to the work of B. Mitterwallner et al. [2]
[1] Lepro et al. , in preparation (2020)
[2] https://journals.aps.org/pre/abstract/10.1103/PhysRevE.101.032408)