Embryons desséchés
Script shell for multiple trajectories analysis. 
The evaluation of error bars was made via the analytical computation of the standard deviation of velocities correlated data. For details about the procedure see the 'OUP-class' script, authored by R.G.

Note for the future:
'We must be careful not to confuse data with the abstractions we use to analyze them.'